remove circular fw and blanket , and move common ivc fns into a separate file
